fgf
gfgfg

+++ app/src/main/java/com/example/moish/aplication_2_forCarRent/model/datasource/MySQL_DBManager.java
@@ -179,7 +179,7 @@ public class MySQL_DBManager implements DB_manager {
     }
 
     @Override
-    public int updateCar( ContentValues values) {
+    public long updateCar(ContentValues values) {
 
 
         try {
@@ -274,6 +274,12 @@ public class MySQL_DBManager implements DB_manager {
     @Override
     public boolean updateCarReserve(ContentValues values) {
         return false;
+
+
+
+
+
+
     }
 
     @Override
